Fix memset
Due to uint32 casting, memset could overwrite 3 bytes past the end

+++ include/string.h
@@ -2,7 +2,7 @@
 #define STRING_H_INCLUDED
 
 void memset32(void *p, int32_t value, uint32_t size);
-void memset(void *p, int32_t value, uint32_t size);
+void memset(void *p, uint8_t value, uint32_t size);
 void memcpy(void *dst, void *src, int n);
 void itoa(int n, char buf[]);
 void uitoa(uint32_t n, char buf[]);
